requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.0.5
pytest==7.4.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
pytz==2023.3 